moved slab metadata and slab address struct to different files
